# pkg/api/v1/metadata.go
package v1

import (
	"time"
)

// Metadata represents the metadata for any resource
//
// swagger:model v1-Metadata
type Metadata struct {
	// UID is the unique in time and space value for this object. It is typically generated by the server on successful creation of a resource and is not allowed to change on PUT operations.
	// Populated by the system.
	//
	// readonly: true
	// example: 01FGTWSBE2RXS960T1TRN6W9D8
	UID string `json:"uid"`

	// CreatedAt is a timestamp representing the server time when this object was created. Clients can not set this value. It is represented in RFC3339 form and is in UTC.
	// Populated by the system.
	//
	// readonly: true
	// example: 2021-09-30T10:22:58.754407751+02:00
	CreatedAt time.Time `json:"created_at" readonly:"true" `

	// ModifiedAt is a timestamp representing the server time when this object was modified. Clients can not set this value. It is represented in RFC3339 form and is in UTC.
	// Populated by the system.
	//
	// readonly: true
	// example: 2021-09-30T10:22:58.754407751+02:00
	ModifiedAt time.Time `json:"modified_at"`
}