# pkg/api/v1/problem.go
package v1

// Problem is an `Problem` implementation
//
// swagger:model v1-Problem
type Problem struct {
	// Type is a URI reference [RFC3986] that identifies the problem type.
	//
	// example: about:default
	Type string `json:"type"`

	// Title is a short, human-readable summary of the problem type.
	//
	// example: Unprocessable Entity
	Title string `json:"title,omitempty"`

	// Status The HTTP status code generated by the origin server for this occurrence of the problem.
	//
	// example: 422
	Status string `json:"status,omitempty"`

	// Detail is a human-readable explanation specific to this occurrence of the problem.
	//
	// example: Your request parameters didn't validate.
	Detail string `json:"detail,omitempty"`

	// Instance is a URI reference that identifies the specific occurrence of the problem.
	//
	// example: /projects/01FGTWSBE2RXS960T1TRN6W9D8
	Instance string `json:"instance,omitempty"`
}